macbook-code/web-search/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiofiles=24.1.0=pypi_0
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.11=pypi_0
aiosignal=1.3.2=pypi_0
aiosqlite=0.20.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
appnope=0.1.4=pyhd8ed1ab_1
asttokens=3.0.0=pyhd8ed1ab_1
async-timeout=5.0.1=pypi_0
attrs=24.3.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
beautifulsoup4=4.12.3=pypi_0
ca-certificates=2024.12.14=hf0a4a13_0
certifi=2024.12.14=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pyhd8ed1ab_1
crawl4ai=0.4.247=pypi_0
cryptography=44.0.0=pypi_0
debugpy=1.8.11=py39h313beb8_0
decorator=5.1.1=pyhd8ed1ab_1
distro=1.9.0=pypi_0
entrypoints=0.4=pyhd8ed1ab_1
exceptiongroup=1.2.2=pypi_0
executing=2.1.0=pyhd8ed1ab_1
fake-http-header=0.3.5=pypi_0
filelock=3.16.1=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.12.0=pypi_0
greenlet=3.1.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.27.2=pypi_0
huggingface-hub=0.27.1=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pypi_0
iniconfig=2.0.0=pypi_0
ipykernel=6.29.5=pyh57ce528_0
ipython=8.12.0=pyhd1c38e8_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.5=pypi_0
jiter=0.8.2=pypi_0
joblib=1.4.2=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh31011fe_1
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
libsodium=1.0.18=h27ca646_1
litellm=1.58.1=pypi_0
lxml=5.3.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mockito=1.5.3=pypi_0
multidict=6.1.0=pypi_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
nltk=3.9.1=pypi_0
numpy=1.24.3=pypi_0
openai=1.59.7=pypi_0
openssl=3.4.0=h81ee809_1
packaging=24.2=pyhd8ed1ab_2
parso=0.8.4=pyhd8ed1ab_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=10.4.0=pypi_0
pip=24.2=py39hca03da5_0
platformdirs=4.3.6=pyhd8ed1ab_1
playwright=1.49.1=pypi_0
pluggy=1.5.0=pypi_0
prompt-toolkit=3.0.48=pyha770c72_1
prompt_toolkit=3.0.48=hd8ed1ab_1
propcache=0.2.1=pypi_0
psutil=6.1.1=pypi_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pycparser=2.22=pypi_0
pydantic=2.10.5=pypi_0
pydantic-core=2.27.2=pypi_0
pyee=12.0.0=pypi_0
pygments=2.19.1=pyhd8ed1ab_0
pyopenssl=25.0.0=pypi_0
pytest=8.3.4=pypi_0
pytest-mockito=0.0.4=pypi_0
python=3.9.21=hb885b13_1
python-dateutil=2.9.0.post0=pyhff2d567_1
python-dotenv=1.0.1=pypi_0
python_abi=3.9=2_cp39
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=py39h313beb8_0
rank-bm25=0.2.2=pypi_0
readline=8.2=h1a28f6b_0
referencing=0.35.1=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rpds-py=0.22.3=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.10.1=pypi_0
setuptools=75.1.0=py39hca03da5_0
six=1.17.0=pyhd8ed1ab_0
sniffio=1.3.1=pypi_0
snowballstemmer=2.2.0=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h80987f9_0
stack_data=0.6.3=pyhd8ed1ab_1
tf-playwright-stealth=1.1.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tiktoken=0.8.0=pypi_0
tk=8.6.14=h6ba3021_0
tokenizers=0.21.0=pypi_0
tomli=2.2.1=pypi_0
tornado=6.1=py39h5161555_1
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
tzdata=2024b=h04d1e81_0
urllib3=2.3.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.44.0=py39hca03da5_0
xxhash=3.5.0=pypi_0
xz=5.4.6=h80987f9_1
yarl=1.18.3=pypi_0
zeromq=4.3.5=h313beb8_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h18a0788_1
